Started adding content
Got rid of the basic angular examples that are automatically added when you make a new app. Added necessary imports and styles for material and mat-icons. Added main toolbar for the app, clickable menu icon, clickable settings icon, and text for the features that will be shown on the main page as a reminder for what I still need to add to the main page. Colors and styles will likely be changed as we go on but I thought blue is a nice color for now